Move icons to assets folder
