+++ src/components/SideBar/SideBar.vue
@@ -17,7 +17,7 @@
 <script>
   import SideBarMenuItem from '@/components/SideBar/SideBarMenuItem'
   import { reloadRouter } from '@/router'
-  import LogoutIcon from '@/components/SideBar/icons/sign_out.svg'
+  import LogoutIcon from '@/assets/icons/sign_out.svg'
 
   export default {
     props: [
